Updated structure of packages - "App" needs to be 1 level above rest to run properly. Updated student.html View.
