+++ src/main/java/Application/Controllers/AdminController.java
@@ -1,7 +1,7 @@
-package Controllers;
+package Application.Controllers;
 
-import DataModel.Applicant;
-import Services.ApplicantService;
+import Application.DataModel.Applicant;
+import Application.Services.ApplicantService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -10,14 +10,18 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/Prof")
-public class ProfessorController {
+@RequestMapping("/Admin")
+public class AdminController {
     @Autowired
     private ApplicantService service;
 
     @GetMapping
     public String applicantForm(Model model, @ModelAttribute Applicant applicant){
         model.addAttribute("applicant", new Applicant());
-        return "prof";
+        return "student";
     }
+
+
+
+
 }
