+++ src/main/java/Application/Controllers/ApplicantController.java
@@ -1,27 +1,21 @@
-package Controllers;
+package Application.Controllers;
 
-import DataModel.Applicant;
-import Services.ApplicantService;
+import Application.DataModel.Applicant;
+import Application.Services.ApplicantService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/Admin")
-public class AdminController {
+public class ApplicantController {
     @Autowired
     private ApplicantService service;
 
-    @GetMapping
-    public String applicantForm(Model model, @ModelAttribute Applicant applicant){
+    @RequestMapping(value = "/applicant")
+    public String index(Model model, @ModelAttribute Applicant applicant) {
         model.addAttribute("applicant", new Applicant());
         return "student";
     }
-
-
-
-
 }
