+++ src/main/java/Application/DataModel/Admin.java
@@ -1,9 +1,9 @@
-package DataModel;
+package Application.DataModel;
 
 import javax.persistence.Entity;
 import javax.persistence.Table;
 
 @Table
 @Entity
-public class Guest extends User {
+public class Admin extends User {
 }
