+++ src/main/java/Application/DataModel/ApplicantRepo.java
@@ -1,4 +1,4 @@
-package DataModel;
+package Application.DataModel;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
