+++ src/main/java/Application/DataModel/Guest.java
@@ -1,9 +1,9 @@
-package DataModel;
+package Application.DataModel;
 
 import javax.persistence.Entity;
 import javax.persistence.Table;
 
 @Table
 @Entity
-public class Admin extends User {
+public class Guest extends User {
 }
