+++ src/main/resources/application.properties
@@ -1,1 +1,4 @@
+test.message: Mkyong
+
+spring.thymeleaf.cache=false
