+++ src/main/resources/templates/login.html
@@ -1,4 +1,4 @@
-
+<!--
 <!DOCTYPE HTML>
 <html xmlns:th="https://www.thymeleaf.org">
