+++ src/main/resources/templates/student.html
@@ -1,30 +1,21 @@
 <!DOCTYPE HTML>
-<html xmlns:th="https://thymeleaf.org">
-
+<html xmlns:th="https://www.thymeleaf.org">
 <head>
-    <title>GRAD APPLICATION</title>
+    <title>Getting Started: Handling Form Submission</title>
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
 </head>
 <body>
-
 <h2>Graduate Application Form</h2>
 <h3>Applicant Profile</h3>
 
+<form action="#" th:action="@{/student.html}" th:object="${applicant}" method="post">
+    <p>Name: <input type="text" th:field="*{name}" /></p>
+    <p>Id: <input type="text" th:field="*{id}" /></p>
+    <p>Email: <input type="text" th:field="*{email}" /></p>
+    <p>Address: <input type="text" th:field="*{address}" /></p>
+    <p>Skills: <input type="text" th:field="*{skillsList}" /></p>
 
-<form action="#" th:action="@{/Applicant}" th:object="${Applicant}" method="post">
-
-    <!--<p>First Name: <input type="text" th:field="${name}" /></p> -->
-
-    <p>Enter Name:<p>
-    <p><input type="text" th:feild="*{name}" /></p>
-    <p>Enter Email Address: <input type="text" th:feild="{*email}" /></p>
-    <p>Enter Professor of choice: <input type="text" /></p>
-    <p>Enter Research field: <input type="text" /></p>
-
-    <p>Upload CV/Resume: <input type="file" id="resumeFile"> <p>
-    <p>Upload Audit: <input type="file" id="auditFile"> <p>
-    <p><input type="submit" value="Submit" />
-        <input type="reset" value="Reset" /></p>
+    <p><input type="submit" value="Submit" /> <input type="reset" value="Reset" /></p>
 </form>
 </body>
 </html>
